backend/app/api/routes/auth.py
from __future__ import annotations

from fastapi import APIRouter, HTTPException, status
from pydantic import field_validator
from sqlmodel import SQLModel

from app.core.config import settings
from app.core.security import create_access_token, verify_password

router = APIRouter(prefix="/api/auth", tags=["auth"])


class TokenRequest(SQLModel):
    password: str

    @field_validator("password", mode="before")
    @classmethod
    def validate_password(cls, value: str) -> str:
        if not isinstance(value, str) or not value:
            raise ValueError("Password cannot be empty")
        return value


class TokenResponse(SQLModel):
    access_token: str
    token_type: str = "bearer"
    expires_in: int


@router.post("/token", response_model=TokenResponse)
def issue_token(payload: TokenRequest) -> TokenResponse:
    if not settings.app_password or not settings.jwt_secret:
        raise HTTPException(
            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
            detail="Auth is not configured",
        )

    if not verify_password(payload.password):
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid password",
            headers={"WWW-Authenticate": "Bearer"},
        )

    return TokenResponse(
        access_token=create_access_token({"sub": "operator"}),
        expires_in=settings.access_token_expire_minutes * 60,
    )
