backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
alembic==1.14.0
psycopg[binary]==3.2.13
pydantic-settings==2.6.1
arq==0.26.1
redis==5.2.0
httpx==0.28.1
dnspython==2.7.0
PyJWT==2.10.1
